+++ include/dpp/coro/awaitable.h
@@ -21,19 +21,27 @@
 
 #pragma once
 
+#include <dpp/export.h>
+
+#if !DPP_BUILD_MODULES
+
 #include <iostream>
 
 #include <dpp/utility.h>
 
+#endif
+
 namespace dpp {
 
-struct awaitable_dummy {
+DPP_EXPORT struct awaitable_dummy {
 	int *promise_dummy = nullptr;
 };
 
 }
 
 #ifndef DPP_NO_CORO
+
+#if !DPP_BUILD_MODULES
 
 #include <dpp/coro/coro.h>
 
@@ -46,6 +54,8 @@ struct awaitable_dummy {
 #include <atomic>
 #include <condition_variable>
 #include <cstdint>
+
+#endif
 
 namespace dpp {
 
@@ -105,7 +115,7 @@ void spawn_sync_wait_job(auto* awaitable, std::condition_variable &cv, auto&& re
 
 } /* namespace detail::promise */
 
-template <typename Derived>
+DPP_EXPORT template <typename Derived>
 class basic_awaitable {
 protected:
 	/**
@@ -204,7 +214,7 @@ public:
  * @tparam T Type of the asynchronous value
  * @see promise
  */
-template <typename T>
+DPP_EXPORT template <typename T>
 class awaitable : public basic_awaitable<awaitable<T>> {
 protected:
 	friend class detail::promise::promise_base<T>;
@@ -507,7 +517,7 @@ public:
  * @tparam T Type of the asynchronous value
  * @see awaitable
  */
-template <typename T>
+DPP_EXPORT template <typename T>
 class basic_promise : public detail::promise::promise_base<T> {
 public:
 	using detail::promise::promise_base<T>::promise_base;
@@ -578,7 +588,7 @@ public:
  *
  * @see awaitable
  */
-template <typename T>
+DPP_EXPORT template <typename T>
 class moveable_promise {
 	/**
 	 * @brief Shared state, wrapped in a unique_ptr to allow move without disturbing an awaitable's promise pointer.
@@ -634,10 +644,10 @@ public:
 	}
 };
 
-template <typename T>
+DPP_EXPORT template <typename T>
 using promise = moveable_promise<T>;
 
-template <typename T>
+DPP_EXPORT template <typename T>
 auto awaitable<T>::abandon() -> uint8_t {
 	uint8_t previous_state = state_flags::sf_broken;
 	if (state_ptr) {
@@ -647,17 +657,17 @@ auto awaitable<T>::abandon() -> uint8_t {
 	return previous_state;
 }
 
-template <typename T>
+DPP_EXPORT template <typename T>
 awaitable<T>::~awaitable() {
 	if_this_causes_an_invalid_read_your_promise_was_destroyed_before_your_awaitable____check_your_promise_lifetime();
 }
 
-template <typename T>
+DPP_EXPORT template <typename T>
 bool awaitable<T>::valid() const noexcept {
 	return state_ptr != nullptr;
 }
 
-template <typename T>
+DPP_EXPORT template <typename T>
 bool awaitable<T>::await_ready() const {
 	if (!this->valid()) {
 		throw dpp::logic_exception("cannot co_await an empty awaitable");
@@ -666,7 +676,7 @@ bool awaitable<T>::await_ready() const {
 	return state & detail::promise::sf_ready;
 }
 
-template <typename T>
+DPP_EXPORT template <typename T>
 template <typename Derived>
 bool awaitable<T>::awaiter<Derived>::await_suspend(detail::std_coroutine::coroutine_handle<> handle) {
 	auto &promise = *awaitable_obj.state_ptr;
@@ -679,7 +689,7 @@ bool awaitable<T>::awaiter<Derived>::await_suspend(detail::std_coroutine::corout
 	return !(previous_flags & detail::promise::sf_ready);
 }
 
-template <typename T>
+DPP_EXPORT template <typename T>
 template <typename Derived>
 T awaitable<T>::awaiter<Derived>::await_resume() {
 	auto &promise = *awaitable_obj.state_ptr;
@@ -697,7 +707,7 @@ T awaitable<T>::awaiter<Derived>::await_resume() {
 
 
 
-template <typename T>
+DPP_EXPORT template <typename T>
 template <typename Derived>
 bool awaitable<T>::awaiter<Derived>::await_ready() const {
 	return static_cast<Derived>(awaitable_obj).await_ready();
@@ -705,31 +715,10 @@ bool awaitable<T>::awaiter<Derived>::await_ready() const {
 
 }
 
+#if !DPP_BUILD_MODULES
+
 #include <dpp/coro/job.h>
 
-namespace dpp {
-
-namespace detail::promise {
-
-template <typename T>
-void spawn_sync_wait_job(auto* awaitable, std::condition_variable &cv, auto&& result) {
-	[](auto* awaitable_, std::condition_variable &cv_, auto&& result_) -> dpp::job {
-		try {
-			if constexpr (std::is_void_v<T>) {
-				co_await *awaitable_;
-				result_.template emplace<1>();
-			} else {
-				result_.template emplace<1>(co_await *awaitable_);
-			}
-		} catch (...) {
-			result_.template emplace<2>(std::current_exception());
-		}
-		cv_.notify_all();
-	}(awaitable, cv, std::forward<decltype(result)>(result));
-}
-
-}
-
-}
+#endif
 
 #endif /* DPP_NO_CORO */
