+++ include/dpp/dpp_fwd.h
@@ -5,5 +5,6 @@ namespace dpp
 
 class cluster;
 struct confirmation_callback_t;
+template <typename T> class event_router_t;
 
 }
